move around some doc block classes
